requirements.txt

numpy==1.21.1
bdd100k==1.0.1
channels==4.1.0
cupy==13.3.0
Django==4.2.16
djangorestframework==3.15.2
GPUtil==1.4.0
imgaug==0.4.0
lap==0.4.0
matplotlib==3.4.3
motmetrics==1.4.0
numba==0.60.0
onnx_graphsurgeon==0.5.2
onnxruntime==1.19.2
onnxsim==0.4.36
opencv_python==4.5.5.64
openvino==2024.4.0
paddle2onnx==1.2.9
paddle_serving_app==0.9.0
paddle_serving_client==0.9.0
paddle_serving_server==0.9.0
paddlepaddle_gpu==3.0.0b1
paddleslim==2.6.0
pandas==1.3.5
Pillow==10.4.0
psutil==5.9.0
psutil==6.0.0
PyCameraList==1.0.0
PyCameraList==1.0.0
pyclipper==1.3.0.post5
pycocotools==2.0.8
pycuda==2024.1.2
PyMySQL==1.1.1
pynvml==11.5.3
PyYAML==6.0.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
sahi==0.11.18
scalabel==0.3.1
scikit_learn==1.5.2
scipy==1.8.0
Shapely==2.0.6
skimage==0.0
tensorrt==10.5.0
tqdm==4.66.5
tritonclient==2.50.0
